pyproject.toml
[project]
name = "tippelaget"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "<4.0,>=3.10"
dependencies = [
    "streamlit (>=1.48.1,<2.0.0)",
    "cognite-sdk (>=7.80.2,<8.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "altair (>=5.5.0,<6.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "openai (>=1.100.1,<2.0.0)"
]

[project.optional-dependencies]
web = [
    "fastapi (>=0.115.0,<1.0.0)",
    "uvicorn[standard] (>=0.32.0,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pydantic-settings (>=2.6.0,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

